Split App runtime hooks
